+++ src/components/stop-summary.tsx
@@ -95,7 +95,7 @@ export function StopSummary({
     'shrink-0 rounded bg-blue-100 px-1.5 py-0.5 text-blue-700 dark:bg-blue-900 dark:text-blue-300';
   const inaccessibleClass =
     'shrink-0 rounded bg-gray-100 px-1.5 py-0.5 text-gray-700 opacity-30 dark:bg-gray-700 dark:text-gray-300';
-  const resolvedAgencyBadgeSize = agencyBadgeSize ?? 'sm';
+  const resolvedAgencyBadgeSize = agencyBadgeSize ?? 'md';
   const resolvedRouteBadgeSize = routeBadgeSize ?? 'xs';
   const wheelchairAccessibleLabel = t('stop.accessibility.wheelchairAccessible');
   const wheelchairNotAccessibleLabel = t('stop.accessibility.wheelchairNotAccessible');
@@ -142,11 +142,12 @@ export function StopSummary({
             <AgencyBadge
               key={agency.agency_id}
               agency={agency}
+              size={resolvedAgencyBadgeSize}
               dataLang={dataLang}
               agencyLangs={resolveAgencyLang(agencies, agency.agency_id)}
               infoLevel={infoLevel}
-              size={resolvedAgencyBadgeSize}
-              disableVerbose
+              // showBorder={false}
+              showBorder={true}
             />
           ))}
       </div>
@@ -160,7 +161,6 @@ export function StopSummary({
               agencyLangs={resolveAgencyLang(agencies, route.agency_id)}
               infoLevel={infoLevel}
               size={resolvedRouteBadgeSize}
-              disableVerbose
               showBorder={true}
             />
           ))}